Modified models to use helper
